Moved clientside config to `...jei.core.config`
