+++ Core/src/main/java/mezz/jei/core/config/IIngredientFilterConfig.java
@@ -1,4 +1,4 @@
-package mezz.jei.gui.config;
+package mezz.jei.core.config;
 
 import mezz.jei.core.search.SearchMode;
