+++ Core/src/main/java/mezz/jei/core/util/gui/HorizontalAlignment.java
@@ -1,4 +1,4 @@
-package mezz.jei.gui.util;
+package mezz.jei.core.util.gui;
 
 public enum HorizontalAlignment {
 	LEFT, CENTER, RIGHT
