+++ Core/src/main/java/mezz/jei/core/util/gui/NavigationVisibility.java
@@ -1,4 +1,4 @@
-package mezz.jei.gui.overlay.options;
+package mezz.jei.core.util.gui;
 
 public enum NavigationVisibility {
 	ENABLED, AUTO_HIDE, DISABLED
