+++ Core/src/main/java/mezz/jei/core/util/gui/VerticalAlignment.java
@@ -1,4 +1,4 @@
-package mezz.jei.gui.util;
+package mezz.jei.core.util.gui;
 
 public enum VerticalAlignment {
 	TOP, CENTER, BOTTOM
