+++ Forge/src/main/java/mezz/jei/forge/JustEnoughItemsClient.java
@@ -2,6 +2,8 @@ package mezz.jei.forge;
 
 import mezz.jei.api.IModPlugin;
 import mezz.jei.common.Internal;
+import mezz.jei.forge.config.ServerConfig;
+import mezz.jei.forge.config.ClientConfig;
 import mezz.jei.gui.config.InternalKeyMappings;
 import mezz.jei.common.gui.textures.JeiSpriteUploader;
 import mezz.jei.common.gui.textures.Textures;
@@ -20,6 +22,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.texture.TextureManager;
 import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
 import net.minecraftforge.client.event.RegisterKeyMappingsEvent;
+import net.minecraftforge.fml.ModLoadingContext;
 
 import java.util.HashSet;
 import java.util.List;
@@ -37,7 +40,11 @@ public class JustEnoughItemsClient {
 	}
 
 	public void register() {
+		ModLoadingContext modLoadingContext = ModLoadingContext.get();
+		/*IServerConfig serverConfig = */ClientConfig.register(modLoadingContext);
+
 		subscriptions.register(RegisterClientReloadListenersEvent.class, this::onRegisterReloadListenerEvent);
+
 	}
 
 	private void onRegisterReloadListenerEvent(RegisterClientReloadListenersEvent event) {
