+++ Forge/src/test/java/mezz/jei/test/lib/TestClientConfig.java
@@ -2,7 +2,7 @@ package mezz.jei.test.lib;
 
 import mezz.jei.core.config.GiveMode;
 import mezz.jei.core.config.IngredientSortStage;
-import mezz.jei.gui.config.IClientConfig;
+import mezz.jei.core.config.IClientConfig;
 
 import java.util.List;
