+++ Forge/src/test/java/mezz/jei/test/lib/TestIngredientFilterConfig.java
@@ -1,7 +1,7 @@
 package mezz.jei.test.lib;
 
 import mezz.jei.core.search.SearchMode;
-import mezz.jei.gui.config.IIngredientFilterConfig;
+import mezz.jei.core.config.IIngredientFilterConfig;
 
 public class TestIngredientFilterConfig implements IIngredientFilterConfig {
 	@Override
