+++ Gui/src/main/java/mezz/jei/gui/ingredients/IngredientFilter.java
@@ -10,8 +10,8 @@ import mezz.jei.api.runtime.IIngredientManager;
 import mezz.jei.api.runtime.IIngredientVisibility;
 import mezz.jei.common.config.DebugConfig;
 import mezz.jei.common.util.Translator;
-import mezz.jei.gui.config.IClientConfig;
-import mezz.jei.gui.config.IIngredientFilterConfig;
+import mezz.jei.core.config.IClientConfig;
+import mezz.jei.core.config.IIngredientFilterConfig;
 import mezz.jei.gui.filter.IFilterTextSource;
 import mezz.jei.gui.overlay.IIngredientGridSource;
 import mezz.jei.gui.search.ElementPrefixParser;
