+++ Gui/src/main/java/mezz/jei/gui/ingredients/IngredientInformationUtil.java
@@ -2,7 +2,7 @@ package mezz.jei.gui.ingredients;
 
 import mezz.jei.api.ingredients.IIngredientHelper;
 import mezz.jei.api.ingredients.IIngredientRenderer;
-import mezz.jei.gui.config.IIngredientFilterConfig;
+import mezz.jei.core.config.IIngredientFilterConfig;
 import mezz.jei.common.util.Translator;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
