+++ Gui/src/main/java/mezz/jei/gui/ingredients/IngredientSorter.java
@@ -2,7 +2,7 @@ package mezz.jei.gui.ingredients;
 
 import mezz.jei.api.runtime.IIngredientManager;
 import mezz.jei.core.config.IngredientSortStage;
-import mezz.jei.gui.config.IClientConfig;
+import mezz.jei.core.config.IClientConfig;
 import mezz.jei.gui.config.IngredientTypeSortingConfig;
 import mezz.jei.gui.config.ModNameSortingConfig;
