+++ Gui/src/main/java/mezz/jei/gui/input/handlers/DeleteItemInputHandler.java
@@ -12,7 +12,7 @@ import mezz.jei.common.network.packets.PacketJei;
 import mezz.jei.gui.util.CheatUtil;
 import mezz.jei.common.util.ServerCommandUtil;
 import mezz.jei.core.config.GiveMode;
-import mezz.jei.gui.config.IClientConfig;
+import mezz.jei.core.config.IClientConfig;
 import mezz.jei.core.config.IWorldConfig;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
