+++ Gui/src/main/java/mezz/jei/gui/overlay/IngredientGridWithNavigation.java
@@ -12,8 +12,8 @@ import mezz.jei.common.util.ImmutableRect2i;
 import mezz.jei.common.util.MathUtil;
 import mezz.jei.core.config.IWorldConfig;
 import mezz.jei.gui.PageNavigation;
-import mezz.jei.gui.config.IClientConfig;
-import mezz.jei.gui.config.IIngredientGridConfig;
+import mezz.jei.core.config.IClientConfig;
+import mezz.jei.core.config.IIngredientGridConfig;
 import mezz.jei.gui.input.IPaged;
 import mezz.jei.gui.input.IRecipeFocusSource;
 import mezz.jei.gui.input.IUserInputHandler;
