+++ Gui/src/main/java/mezz/jei/gui/overlay/bookmarks/BookmarkOverlay.java
@@ -13,7 +13,7 @@ import mezz.jei.common.network.IConnectionToServer;
 import mezz.jei.common.util.ImmutableRect2i;
 import mezz.jei.core.config.IWorldConfig;
 import mezz.jei.gui.bookmarks.BookmarkList;
-import mezz.jei.gui.config.IClientConfig;
+import mezz.jei.core.config.IClientConfig;
 import mezz.jei.gui.elements.GuiIconToggleButton;
 import mezz.jei.gui.input.IRecipeFocusSource;
 import mezz.jei.gui.input.IUserInputHandler;
