--- Gui/src/main/java/mezz/jei/gui/overlay/options/package-info.java
@@ -1,7 +1,0 @@
-@ParametersAreNonnullByDefault
-@MethodsReturnNonnullByDefault
-package mezz.jei.gui.overlay.options;
-
-import net.minecraft.MethodsReturnNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
