+++ Gui/src/main/java/mezz/jei/gui/search/ElementPrefixParser.java
@@ -9,7 +9,7 @@ import mezz.jei.core.search.LimitedStringStorage;
 import mezz.jei.core.search.PrefixInfo;
 import mezz.jei.core.search.SearchMode;
 import mezz.jei.core.search.suffixtree.GeneralizedSuffixTree;
-import mezz.jei.gui.config.IIngredientFilterConfig;
+import mezz.jei.core.config.IIngredientFilterConfig;
 import mezz.jei.gui.ingredients.IListElementInfo;
 
 import java.util.ArrayList;
