+++ Gui/src/main/java/mezz/jei/gui/startup/JeiGuiStarter.java
@@ -21,10 +21,10 @@ import mezz.jei.core.config.IWorldConfig;
 import mezz.jei.core.config.file.FileWatcher;
 import mezz.jei.gui.bookmarks.BookmarkList;
 import mezz.jei.gui.config.IBookmarkConfig;
-import mezz.jei.gui.config.IClientConfig;
-import mezz.jei.gui.config.IIngredientFilterConfig;
-import mezz.jei.gui.config.IIngredientGridConfig;
-import mezz.jei.gui.config.IJeiClientConfigs;
+import mezz.jei.core.config.IClientConfig;
+import mezz.jei.core.config.IIngredientFilterConfig;
+import mezz.jei.core.config.IIngredientGridConfig;
+import mezz.jei.core.config.IJeiClientConfigs;
 import mezz.jei.gui.config.IngredientTypeSortingConfig;
 import mezz.jei.gui.config.ModNameSortingConfig;
 import mezz.jei.gui.events.GuiEventHandler;
