+++ Gui/src/main/java/mezz/jei/gui/startup/OverlayHelper.java
@@ -11,9 +11,9 @@ import mezz.jei.common.input.IInternalKeyMappings;
 import mezz.jei.common.network.IConnectionToServer;
 import mezz.jei.core.config.IWorldConfig;
 import mezz.jei.gui.bookmarks.BookmarkList;
-import mezz.jei.gui.config.IClientConfig;
-import mezz.jei.gui.config.IIngredientFilterConfig;
-import mezz.jei.gui.config.IIngredientGridConfig;
+import mezz.jei.core.config.IClientConfig;
+import mezz.jei.core.config.IIngredientFilterConfig;
+import mezz.jei.core.config.IIngredientGridConfig;
 import mezz.jei.gui.filter.IFilterTextSource;
 import mezz.jei.gui.overlay.IIngredientGridSource;
 import mezz.jei.gui.overlay.IngredientGrid;
